Starting to retire the SArray class. The attributes of being fixed and being discrete valued are now properties of the node, not its data member.
The SArray::isDiscreteValued function is kept for the moment for the
purposes of reading/writing the data table.

The extra argument "absolute" is removed from the l and u member functions
of the DistScalar class. The required functionality (for MixCompiler)
is supplied by a new member function isSupportFixed that takes a boolean mask.